Move the process primitives to platform/process.rs
Bounded child execution, the python/pip capture helpers, env isolation,
window suppression and the Windows job object and console signal plumbing,
plus their tests. tail_for_log, run_python_capture_bounded and
pip_install_blocking become pub(super) for their callers still in mod.rs;
the externally used entry points are re-exported so no call site changes.
JOB_HELPER_TEST follows the job tests to their new module path.
